NOT WORKING JWT

+++ database/mongo/delete.go
@@ -9,11 +9,11 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
-func Delete(url string) (*mongo.DeleteResult, error) {
+func Delete(url, owner string) (*mongo.DeleteResult, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 
-	r, err := Coll.DeleteOne(ctx,  bson.M{"_id": url})
+	r, err := Coll.DeleteOne(ctx, bson.M{"_id": url, "owner": owner})
 	if err != nil {
 		fmt.Print(err)
 		return nil, err
